A lot of changes to how the client functions.
The for loop of the API endpoints was removed for favour for calling
HealthGraph#apiCall with a opts.resource. The apiCall function takes an
object so we can pass parameters to the url and make it more flexible
when it comes to access_keys and more.

Added a way to request an authorization_code from the API.

Added a function to get the activity feed for the user. It takes a
get_next opts param which will continue to hit the API until it reaches
the last page of the feed.